normalize line endings for linux

+++ Events/Guild/guildDelete.js
@@ -8,10 +8,15 @@ module.exports = {
         // remove all guild related data from the database
         const dbTablesObject = JSON.parse(fs.readFileSync('./objects/database-default-tables.json'));
         let tableNames = dbTablesObject["table_names"];
-        tableNames.forEach(async (table) => {
+        
+        for(const table of tableNames) {
+            if(table == "botconfig")
+                continue;
             try{
                 await poolConnection.query(`DELETE FROM ${table} WHERE guild=$1`, [guild.id]);
-            } catch(err) { console.error(`guildDelete failed removing guild from ${table} table`) };
-        });
+            } catch(err) { console.error(`guildDelete failed removing guild from ${table} table\n${err}`) };
+
+        }
+
     }
 }
